# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: cdCTF API
  version: 0.1.0
  description: cdCTF API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with nickname and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-email:
    get:
      operationId: verifyEmail
      tags: [auth]
      summary: Verify email with token
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Email verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # CTF
...
  /competitions/{id}:
    get:
      operationId: getCompetition
      tags: [competitions]
      summary: Get competition detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Competition detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /competitions/{id}/join:
    post:
      operationId: joinCompetition
      tags: [competitions]
      summary: Join a competition
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "403":
          description: Cannot join
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /competitions/{id}/scoreboard:
    get:
      operationId: getCompetitionScoreboard
      tags: [competitions]
      summary: Get competition scoreboard
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Competition scoreboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScoreboardEntry"

  # PROFILE
...
  /admin/competitions:
    post:
      operationId: adminCreateCompetition
      tags: [admin]
      summary: Create a competition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompetitionBody"
      responses:
        "201":
          description: Competition created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Competition"
  /admin/competitions/{id}:
    put:
      operationId: adminUpdateCompetition
      tags: [admin]
      summary: Update a competition
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompetitionBody"
      responses:
        "200":
          description: Competition updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Competition"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # Auth
    RegisterBody:
      type: object
      properties:
        nickname:
          type: string
          minLength: 3
          maxLength: 32
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required: [nickname, email, password]
    LoginBody:
      type: object
      properties:
        nickname:
          type: string
        password:
          type: string
      required: [nickname, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    User:
      type: object
      properties:
        id:
          type: integer
        nickname:
          type: string
        email:
          type: string
        avatarUrl:
          type: string
          nullable: true
        points:
          type: integer
        role:
          type: string
          enum: [user, admin]
        emailVerified:
          type: boolean
        isBlocked:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, nickname, email, points, role, emailVerified, isBlocked, createdAt]

    # CTF
    CtfChallenge:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard, insane]
        points:
          type: integer
        solvedCount:
          type: integer
        isSolved:
          type: boolean
        isBlocked:
          type: boolean
      required: [id, name, category, difficulty, points, solvedCount, isSolved, isBlocked]
...
    ScoreboardEntry:
      type: object
      properties:
        rank:
          type: integer
        userId:
          type: integer
        nickname:
          type: string
        avatarUrl:
          type: string
          nullable: true
        points:
          type: integer
        solvedCtfCount:
          type: integer
        completedLessonsCount:
          type: integer
        titles:
          type: array
          items:
            type: string
      required: [rank, userId, nickname, points, solvedCtfCount, completedLessonsCount, titles]
...
    Competition:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [public, private]
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        status:
          type: string
          enum: [upcoming, active, ended]
        participantCount:
          type: integer
        ctfCount:
          type: integer
        isJoined:
          type: boolean
      required: [id, name, type, startTime, endTime, status, participantCount, ctfCount, isJoined]
    CompetitionDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [public, private]
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        status:
          type: string
          enum: [upcoming, active, ended]
        participantCount:
          type: integer
        isJoined:
          type: boolean
        challenges:
          type: array
          items:
            $ref: "#/components/schemas/CtfChallenge"
        certificateUrl:
          type: string
          nullable: true
      required: [id, name, type, startTime, endTime, status, participantCount, isJoined, challenges]

    # Profile
...
    CreateCompetitionBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [public, private]
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        ctfIds:
          type: array
          items:
            type: integer
      required: [name, type, startTime, endTime, ctfIds]
...